+++ cmd/plugins.go
@@ -20,6 +20,7 @@ import (
 	_ "github.com/tsarna/vinculum/triggers/shutdown"
 	_ "github.com/tsarna/vinculum/triggers/signals"
 	_ "github.com/tsarna/vinculum/triggers/start"
+	_ "github.com/tsarna/vinculum/triggers/watch"
 	_ "github.com/tsarna/vinculum/triggers/watchdog"
 	_ "github.com/tsarna/vinculum/functions"
 )