+++ triggers/watch/testplugins_test.go
@@ -1,0 +1,4 @@
+package watch_test
+
+// Import functions package to trigger init() registrations during testing.
+import _ "github.com/tsarna/vinculum/functions"